Write custom hook for the certificates page
